Refactor more, add dummy test.
