+++ packages/unmock-matcher/package.json
@@ -14,11 +14,7 @@
   },
   "dependencies": {
     "@types/debug": "^4.1.4",
-    "@types/node": "^12.0.8",
-    "@types/readable-stream": "^2.3.1",
-    "content-type": "^1.0.4",
     "debug": "^4.1.1",
-    "readable-stream": "^3.4.0",
     "unmock-core": "file:../unmock-core"
   }
 }
