+++ packages/unmock-matcher/src/__tests__/matcher.test.ts
@@ -1,0 +1,5 @@
+describe("http request matcher", () => {
+  it("works", () => {
+    expect(0).toBe(1);
+  });
+});
