+++ packages/unmock-node/package.json
@@ -13,11 +13,13 @@
     "url": "https://github.com/unmock/unmock-js/issues"
   },
   "dependencies": {
+    "@types/readable-stream": "^2.3.1",
     "expect": "^24.7.1",
     "ini": "^1.3.5",
     "js-yaml": "^3.13.1",
     "mitm": "^1.7.0",
     "mkdirp": "^0.5.1",
+    "readable-stream": "^3.4.0",
     "unmock-core": "file:../unmock-core",
     "unmock-hash": "file:../unmock-hash",
     "winston": "^3.2.1"
